Moving getProductData and the onchangeHandler to Home.jsx for components to work together. Removing exta row in test.
