+++ cypress/integration/userCanSortProductsByCategories.feature.js
@@ -21,7 +21,6 @@ describe("User can see list of", () => {
 
     it("sorted by categories", () => {
       cy.get('[data-cy="dropdown"]').click();
-
       cy.get("[role='option']").contains('Veggie').click();
     });
   });
