+++ src/components/DisplayCategoriesList.jsx
@@ -13,14 +13,6 @@ class DisplayCategoriesList extends Component {
 		options: [],
 	};
 
-	onChangeHandler = async (event) => {
-    let selectedCategory = event.target.textContent.toLowerCase()
-    let response = await axios.get(
-      `/categories/${selectedCategory}`, 
-    ) 
-    debugger
-  };
-
 	render() {
 		return (
 			<Menu compact>
